Modfied Delivery Covarage

+++ app/Repositories/DeliveryCoverageRepository.php
@@ -2,24 +2,42 @@
 
 namespace App\Repositories;
 
+use App\Http\Controllers\Controller;
 use App\Interfaces\DeliveryCoverageInterface;
 use App\Models\deliveryCoverage;
+use Illuminate\Support\Facades\Auth;
 
-class DeliveryCoverageRepository implements DeliveryCoverageInterface
+class DeliveryCoverageRepository extends Controller implements DeliveryCoverageInterface
 {
-    public function deliveryCoverage($details)
+    /**
+     * Undocumented function
+     *
+     * @param [type] $details
+     * @return void
+     */
+    public function storeCoverage($details)
     {
         $coverageArea = deliveryCoverage::create($details);
         return $coverageArea;
     }
-
+    /**
+     * Undocumented function
+     *
+     * @param [type] $shop_id
+     * @return void
+     */
     public function getAllDeliveryCoverage($shop_id)
     {
         $coverageAreas = deliveryCoverage::where('shop_id', $shop_id)->get();
         return $coverageAreas;
     }
 
-
+    /**
+     * Undocumented function
+     *
+     * @param [type] $id
+     * @return void
+     */
     public function getDeliveryCoverage($id)
     {
         $coverageAreas = deliveryCoverage::find($id);
@@ -27,13 +45,34 @@ class DeliveryCoverageRepository implements DeliveryCoverageInterface
     }
 
 
+    /**
+     * Undocumented function
+     *
+     * @param [type] $id
+     * @return void
+     */
     public function deleteDeliveryCoverage($id)
     {
-        $coverageAreas = deliveryCoverage::find($id);
-        $coverageAreas->delete();
+        $shop_id = Auth::user()->providerShopDetails->id;
+        $coverageAreasExists = deliveryCoverage::where('id', $id)->where('shop_id', $shop_id)->exists();
+
+        if ($coverageAreasExists) {
+            $coverageAreas = deliveryCoverage::findOrFail($id);
+            $coverageAreas->delete();
+        }else{
+            return $this->errorResponseWithMessage('the delivery coverage not exists ,or not own this delivery coverage', 401);
+ 
+        }
     }
 
 
+    /**
+     * Undocumented function
+     *
+     * @param [type] $id
+     * @param [type] $data
+     * @return void
+     */
     public function updateDeliveryCoverage($id, $data)
     {
         $delivery = deliveryCoverage::findOrFail($id);
